requirements.txt
Django==5.0.2
django-mptt==0.16.0
djangorestframework==3.14.0
drf-spectacular==0.27.1
drf-spectacular-sidecar==2024.3.1
factory-boy==3.3.0
python-dotenv==1.0.1
